Split MVAcommon into one file for each class
